fix sanitize html (#860)
