+++ composer.json
@@ -22,13 +22,14 @@
   "require": {
     "php": "^7.2",
     "ext-ctype": "*",
+    "ext-dom": "*",
     "ext-iconv": "*",
     "ext-json": "*",
-    "ext-zlib": "*",
-    "ext-dom": "*",
     "ext-libxml": "*",
+    "ext-zlib": "*",
     "doctrine/doctrine-fixtures-bundle": "^2.3",
     "doctrine/doctrine-migrations-bundle": "^1.3",
+    "ezyang/htmlpurifier": "^4.12",
     "facebook/graph-sdk": "^5.4",
     "friendsofsymfony/jsrouting-bundle": "2.2.*",
     "friendsofsymfony/user-bundle": "^2.1",
