+++ src/Enhavo/Bundle/FormBundle/composer.json
@@ -13,7 +13,8 @@
   ],
   "require": {
     "php": "^7.2",
-    "enhavo/app-bundle": "dev-master"
+    "enhavo/app-bundle": "dev-master",
+    "ezyang/htmlpurifier": "^4.12"
   },
   "require-dev": {
     "phpunit/phpunit": "^7.4"
