+++ symfony.lock
@@ -116,6 +116,9 @@
     "evenement/evenement": {
         "version": "v2.1.0"
     },
+    "ezyang/htmlpurifier": {
+        "version": "v4.12.0"
+    },
     "facebook/graph-sdk": {
         "version": "5.6.3"
     },
